requirements.txt
asgiref==3.4.1
astroid==2.9.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==36.0.0
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.46.0
django-countries==7.2.1
django-filter==21.1
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
idna==3.3
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
oauthlib==3.1.1
Pillow==8.4.0
pkg_resources==0.0.0
platformdirs==2.4.0
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pylint==2.12.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.7
wrapt==1.13.3
